// src/lib/types.ts
export type PresentationSlide = {
  title: string;
  content: string[];
  imageUrl?: string;
};

export type PresentationData = {
  title: string;
  agenda: string[];
  slides: PresentationSlide[];
};